Database is now created dynamically if it doesn't exist.
The database.db file is no longer tracked.
Delete database.db. Will be created dynamically the first time the
application runs.
Add create_tables.sql. The SQL script that contains the tables schemas
that are being created in the database.
Add docstrings and some comments.
Add a confirmation dialog when the user wants to exit the application.